requirements.txt
pytest~=8.3.5
playwright~=1.52.0
requests
pytest-html
python-dotenv~=1.1.0
pytest-playwright~=0.7.0